# Define attributes for comment model