--- a/app/Transformers/CommentTransformer.php
+++ b/app/Transformers/CommentTransformer.php
@@ -40,12 +40,12 @@ class CommentTransformer extends TransformerAbstract
 
             'legacy_name' => $comment->legacyName(),
 
-            'created_at' => json_time($comment->created_at),
-            'updated_at' => json_time($comment->updated_at),
+            'created_at' => $comment->created_at_json,
+            'updated_at' => $comment->updated_at_json,
 
-            'deleted_at' => json_time($comment->deleted_at),
+            'deleted_at' => $comment->deleted_at_json,
 
-            'edited_at' => json_time($comment->edited_at),
+            'edited_at' => $comment->edited_at_json,
             'edited_by_id' => $comment->edited_by_id,
         ];
     }